+++ lib/dbus.rb
@@ -11,17 +11,14 @@
 # See the file "COPYING" for the exact licensing terms.
 
 module DBus
-  # Byte signifying big endianness.
+  # Protocol character signifying big endianness.
   BIG_END = "B"
-  # Byte signifying little endianness.
+  # Protocol character signifying little endianness.
   LIL_END = "l"
 
-  # Byte signifying the host's endianness.
-  HOST_END = if [0x01020304].pack("L").unpack1("V") == 0x01020304
-               LIL_END
-             else
-               BIG_END
-             end
+  # Protocol character signifying the host's endianness.
+  # "S": unpack as uint16, native endian
+  HOST_END = { 1 => BIG_END, 256 => LIL_END }.fetch("\x00\x01".unpack1("S"))
 end
 # ^ That's because dbus/message needs HOST_END early
 
@@ -34,17 +31,21 @@ require_relative "dbus/emits_changed_signal"
 require_relative "dbus/error"
 require_relative "dbus/introspect"
 require_relative "dbus/logger"
+require_relative "dbus/main"
 require_relative "dbus/marshall"
 require_relative "dbus/matchrule"
 require_relative "dbus/message"
 require_relative "dbus/message_queue"
+require_relative "dbus/node_tree"
 require_relative "dbus/object"
 require_relative "dbus/object_manager"
 require_relative "dbus/object_path"
+require_relative "dbus/object_server"
 require_relative "dbus/platform"
 require_relative "dbus/proxy_object"
 require_relative "dbus/proxy_object_factory"
 require_relative "dbus/proxy_object_interface"
+require_relative "dbus/proxy_service"
 require_relative "dbus/raw_message"
 require_relative "dbus/type"
 require_relative "dbus/xml"