+++ lib/dbus/object.rb
@@ -26,20 +26,29 @@ module DBus
     my_class_attribute :intfs
     self.intfs = {}
 
-    # The service that the object is exported by.
-    attr_writer :service
+    # @return [Connection] the connection the object is exported by
+    attr_reader :connection
 
     @@cur_intf = nil # Interface
     @@intfs_mutex = Mutex.new
 
     # Create a new object with a given _path_.
-    # Use Service#export to export it.
+    # Use ObjectServer#export to export it.
     def initialize(path)
       @path = path
-      @service = nil
+      @connection = nil
+    end
+
+    # @param connection [Connection] the connection the object is exported by
+    def connection=(connection)
+      @connection = connection
+      # deprecated, keeping @service for compatibility
+      @service = connection&.object_server
     end
 
     # Dispatch a message _msg_ to call exported methods
+    # @param msg [Message] only METHOD_CALLS do something
+    # @api private
     def dispatch(msg)
       case msg.message_type
       when Message::METHOD_CALL
@@ -69,7 +78,7 @@ module DBus
           dbus_msg_exc = msg.annotate_exception(e)
           reply = ErrorMessage.from_exception(dbus_msg_exc).reply_to(msg)
         end
-        @service.bus.message_queue.push(reply)
+        @connection.message_queue.push(reply)
       end
     end
 
@@ -307,7 +316,7 @@ module DBus
     # @param sig [Signal]
     # @param args arguments for the signal
     def emit(intf, sig, *args)
-      @service.bus.emit(@service, self, intf, sig, *args)
+      @connection.emit(nil, self, intf, sig, *args)
     end
 
     # Defines a signal for the object with a given name _sym_ and _prototype_.
@@ -316,7 +325,7 @@ module DBus
 
       cur_intf = @@cur_intf
       signal = Signal.new(sym.to_s).from_prototype(prototype)
-      cur_intf.define(Signal.new(sym.to_s).from_prototype(prototype))
+      cur_intf.define(signal)
 
       # ::Module#define_method(name) { body }
       define_method(sym.to_s) do |*args|