+++ spec/spec_helper.rb
@@ -1,51 +1,13 @@
 # frozen_string_literal: true
 
-coverage = if ENV["COVERAGE"]
-             ENV["COVERAGE"] == "true"
-           else
-             # heuristics: enable for interactive builds (but not in OBS)
-             ENV["DISPLAY"]
-           end
-
-if coverage
-  require "simplecov"
-  SimpleCov.root File.expand_path("..", __dir__)
-
-  # do not cover specs
-  SimpleCov.add_filter "_spec.rb"
-  # do not cover the activesupport helpers
-  SimpleCov.add_filter "/core_ext/"
-  # measure all if/else branches on a line
-  SimpleCov.enable_coverage :branch
-
-  SimpleCov.start
-
-  # additionally use the LCOV format for on-line code coverage reporting at CI
-  if ENV["COVERAGE_LCOV"] == "true"
-    require "simplecov-lcov"
-
-    SimpleCov::Formatter::LcovFormatter.config do |c|
-      c.report_with_single_file = true
-      # this is the default Coveralls GitHub Action location
-      # https://github.com/marketplace/actions/coveralls-github-action
-      c.single_report_path = "coverage/lcov.info"
-    end
-
-    SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter.new [
-      SimpleCov::Formatter::HTMLFormatter,
-      SimpleCov::Formatter::LcovFormatter
-    ]
-  end
-end
+require_relative "coverage_helper"
 
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 
-if Object.const_defined? "RSpec"
-  # http://betterspecs.org/#expect
-  RSpec.configure do |config|
-    config.expect_with :rspec do |c|
-      c.syntax = :expect
-    end
+# http://betterspecs.org/#expect
+RSpec.configure do |config|
+  config.expect_with :rspec do |c|
+    c.syntax = :expect
   end
 end
 
@@ -102,7 +64,7 @@ end
 
 def with_service_by_activation(&block)
   name = "org.ruby.service"
-  exec = "#{TOPDIR}/spec/service_newapi.rb"
+  exec = "#{TOPDIR}/spec/mock-service/spaghetti-monster.rb"
 
   service_dir = "#{ENV["XDG_DATA_DIRS"]}/dbus-1/services"
   FileUtils.mkdir_p service_dir
@@ -118,7 +80,9 @@ def with_service_by_activation(&block)
 
   block.call
 
-  system "pkill -f #{exec}"
+  # This would kill also other instances,
+  # namely on the bus set up by test_env.
+  ## system "pkill -f #{exec}"
 end
 
 # Make a binary string from readable YAML pieces; see data/marshall.yaml